==> Script for Analysis/P2 MergeDataset.ipynb <==
import pandas as pd
import os
from google.colab import drive

drive.mount('/content/drive')

folder_path = '/content/drive/MyDrive/StackOverflow_Data/'

all_files = [os.path.join(folder_path, f) for f in os.listdir(folder_path) if f.endswith('.csv')]
all_files.sort() # Mengurutkan file secara numerik

output_file = '/content/drive/MyDrive/StackOverflow_Data/all_questions_2020-2024.csv'

print("Memulai penggabungan file...")
first_file = True
for f in all_files:
    # Membaca file CSV
    df_temp = pd.read_csv(f)

    if first_file:
        df_temp.to_csv(output_file, mode='w', index=False)
        first_file = False
    else:
        df_temp.to_csv(output_file, mode='a', header=False, index=False)

    print(f"Ditambahkan: {os.path.basename(f)}")

print(f"Penggabungan selesai. Semua data tersimpan di {output_file}")

file_path = '/content/drive/MyDrive/StackOverflow_Data/all_questions_2020-2024.csv'

# Cek apakah file ada di Drive
if os.path.exists(file_path):
    print("File all_questions_2020-2024.csv ditemukan!")

    # Baca file untuk memastikan isinya ada
    df_check = pd.read_csv(file_path, nrows=5)
    print("\nBerikut 5 baris pertama dari file:")
    print(df_check)
else:
    print("File tidak ditemukan di Google Drive.")

file_path = "/content/drive/MyDrive/StackOverflow_Data/all_questions_2020-2024.csv"
df = pd.read_csv(file_path)
