==> Script for Analysis/P3 FilteringData.ipynb <==
import pandas as pd
import os
from google.colab import drive

drive.mount('/content/drive')

file_path = "/content/drive/MyDrive/StackOverflow_Data/all_questions_2020-2024.csv"

chunksize = 100000
chunks = pd.read_csv(file_path, chunksize=chunksize)

row_count = 0
for chunk in chunks:
    row_count += len(chunk)

print("Total baris:", row_count)

file_path = "/content/drive/MyDrive/StackOverflow_Data/all_questions_2020-2024.csv"

output_path = "/content/drive/MyDrive/StackOverflow_Data/stackoverflow_filtered_2020-2024.csv"

chunksize = 100000
chunks = pd.read_csv(file_path, chunksize=chunksize)

open(output_path, "w").close()

for i, chunk in enumerate(chunks):
    filtered_chunk = chunk[["Title", "Body", "Tags", "CreationDate"]]

    if i == 0:
        filtered_chunk.to_csv(output_path, index=False, mode="w")
    else:
        filtered_chunk.to_csv(output_path, index=False, mode="a", header=False)

    print(f"Chunk {i+1} selesai diproses ({len(filtered_chunk)} baris)")

print(f"\nFiltering selesai! Dataset baru disimpan di: {output_path}")
